update main.py file with add middleware and some method change

+++ backend/main.py
@@ -1,9 +1,19 @@
+from http.client import HTTPException
+
 from fastapi import FastAPI
 from pydantic import BaseModel
 import psycopg2
+from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
 
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["http://localhost:3000"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
 # connect PostgreSQL
 conn = psycopg2.connect(
     dbname="todo_db",
@@ -22,13 +32,29 @@ class Task(BaseModel):
 def get_tasks():
     cursor.execute("SELECT * FROM task")
     rows = cursor.fetchall()
-    return rows
 
+    return [
+        {
+            "id": row[0],
+            "name": row[1],
+            "checked": row[2]
+        }
+        for row in rows
+    ]
 # get task by id
 @app.get("/task/{task_id}")
 def get_task(task_id: int):
     cursor.execute("SELECT * FROM task WHERE id=%s", (task_id,))
-    return cursor.fetchone()
+    task = cursor.fetchone()
+
+    if not task:
+        return {"error": "Task not found"}
+
+    return {
+        "id": task[0],
+        "name": task[1],
+        "checked": task[2]
+    }
 
 # add new tasks
 @app.post("/task")
@@ -38,21 +64,43 @@ def add_task(task: Task):
         (task.name, task.checked)
     )
     conn.commit()
-    return {"message": "Task added"}
+    return {
+    "name": task.name,
+    "checked": task.checked
+}
 
 #update task
 @app.put("/task/{task_id}")
 def update_task(task_id: int, task: Task):
+    
+    cursor.execute("SELECT id FROM task WHERE id=%s", (task_id,))
+    existing = cursor.fetchone()
+
+    if not existing:
+        raise HTTPException(status_code=404, detail="Task not found")
+
     cursor.execute(
         "UPDATE task SET name=%s, checked=%s WHERE id=%s",
         (task.name, task.checked, task_id)
     )
     conn.commit()
-    return {"message": "Updated"}
+
+    return {
+        "id": task_id,
+        "name": task.name,
+        "checked": task.checked
+    }
 
 #delete task
 @app.delete("/task/{task_id}")
 def delete_task(task_id: int):
+    cursor.execute("SELECT id FROM task WHERE id=%s", (task_id,))
+    task = cursor.fetchone()
+
+    if not task:
+        raise HTTPException(status_code=404, detail="Task not found")
+
     cursor.execute("DELETE FROM task WHERE id=%s", (task_id,))
     conn.commit()
-    return {"message": "Deleted"}
+
+    return {"message": "Task deleted successfully"}
